+++ visualization/app/codeCharta/state/store/dynamicSettings/colorRange/colorRange.selector.ts
@@ -1,0 +1,4 @@
+import { createSelector } from "../../../angular-redux/createSelector"
+import { dynamicSettingsSelector } from "../dynamicSettings.selector"
+
+export const colorRangeSelector = createSelector([dynamicSettingsSelector], dynamicAppSettings => dynamicAppSettings.colorRange)