+++ visualization/app/codeCharta/ui/colorSettingsPanel/colorSettingsPanel.module.ts
@@ -1,9 +1,11 @@
-import "../../state/state.module"
-
 import angular from "angular"
+import { downgradeComponent } from "@angular/upgrade/static"
 
+import "../../state/state.module"
 import { colorSettingsPanelComponent } from "./colorSettingsPanel.component"
+import { ColorPickerForMapColorComponent } from "../colorPickerForMapColor/colorPickerForMapColor.component"
 
 angular
 	.module("app.codeCharta.ui.colorSettingsPanel", ["app.codeCharta.state"])
 	.component(colorSettingsPanelComponent.selector, colorSettingsPanelComponent)
+	.directive("ccColorPickerForMapColor", downgradeComponent({ component: ColorPickerForMapColorComponent }))