+++ visualization/setupJest.ts
@@ -1,3 +1,4 @@
+/* eslint-disable unicorn/prevent-abbreviations */
 /* eslint-disable @typescript-eslint/triple-slash-reference */
 
 // After removing below if block, and resorting to "import" this manual type reference becomes obsolete
@@ -8,4 +9,9 @@ if (process.env.TEST_TYPE !== "e2e") {
 	require("angular")
 	require("angular-mocks")
 	require("jest-preset-angular/setup-jest")
+}
+
+const numberToLocaleString = Number.prototype.toLocaleString
+Number.prototype.toLocaleString = function (locale = "en-US") {
+	return numberToLocaleString.call(this, locale)
 }