+++ CHANGELOG.md
@@ -7,13 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ## [unreleased] (Added 🚀 | Changed | Removed 🗑 | Fixed 🐞 | Chore 👨‍💻 👩‍💻)
 
+### Changed
+
+-   Replace all color pickers in the context of ongoing Angular migration [#2560](https://github.com/MaibornWolff/codecharta/pull/2560).
+-   Remove suspicious metrics from custom map configurations [#2564](https://github.com/MaibornWolff/codecharta/pull/2564)
+
 ### Fixed 🐞
 
 -   Fix showing names of custom configs properly [#2557](https://github.com/MaibornWolff/codecharta/pull/2557)
 
-### Changed
+### Chore 👨‍💻 👩‍💻
 
--   Remove suspicious metrics from custom map configurations [#2564](https://github.com/MaibornWolff/codecharta/pull/2564)
+-   Migrate legend panel to Angular [#2560](https://github.com/MaibornWolff/codecharta/pull/2560).
 
 ## [1.85.0] - 2021-12-13